+++ src/features/projects/runbooks/runs/index.ts
@@ -1,3 +1,5 @@
 export * from "./createRunbookRunCommandV1";
 export * from "./run-runbook";
+export * from "./runbookRun";
+export * from "./runbookRunRepository";
 export * from "./runbookRunServerTask";
